Feat: reuse functions for generate new point form
